+++ clang/include/clang/Sema/SemaConcept.h
@@ -162,7 +162,7 @@ private:
 
   // Each Literal has a unique value that is enough to establish
   // its identity.
-  // Some constraints (fold expended) requires special subsumption
+  // Some constraints (fold expended) require special subsumption
   // handling logic beyond comparing values, so we store a flag
   // to let us quickly dispatch to each kind of variable.
   struct Literal {
@@ -230,14 +230,17 @@ private:
   DNFFormula DNF(const NormalizedConstraint &C);
 
   template <typename FormulaType>
-  FormulaType Normalize(const NormalizedConstraint &C);
-  void AddClauseToFormula(Formula &F, Clause C);
+  FormulaType
+  Normalize(const NormalizedConstraint &C,
+            NormalizedConstraint::CompoundConstraintKind ParentKind);
+  void AddNonRedundantClauseToFormula(Formula &F, Clause C);
+  void AddUniqueClauseToFormula(Formula &F, Clause C);
   bool IsSuperSet(const Clause &A, const Clause &B);
 
   Literal find(AtomicConstraint *);
   Literal find(FoldExpandedConstraint *);
 
-  uint16_t getNewId();
+  uint16_t getNewLiteralId();
 };
 
 } // clang